Better handling of folders with non-zip files

+++ zipify.sh
@@ -10,6 +10,11 @@ _die ()
   exit 1
 }
 
+_skip ()
+{
+  echo "Skipping: $1"
+}
+
 # find the necessary binaries, unzip and 7z must be in $PATH
 _getbin ()
 {
@@ -23,15 +28,18 @@ _getbin ()
 # extract zip contents and then add them to 7z archive
 _zipify ()
 {
-  echo ======== "$1" ========
-  SZFILE=$(basename "$1" .zip).7z
-  mkdir "$1.dir" && cd "$1.dir"
-  $UNZIP ../"$1"
-  $SZIP a "$SZFILE" ./*
-  mv "$SZFILE" .. 
-  cd .. && rm -rf "$1.dir"
-  if [ -f ./"$SZFILE" ]; then
-    rm "$1"
+  if $(file -i "$1" | grep -q 'application/zip'); then
+    SZFILE=$(basename "$1" .zip).7z
+    mkdir "$1.dir" && cd "$1.dir"
+    $UNZIP ../"$1"
+    $SZIP a "$SZFILE" ./*
+    mv "$SZFILE" .. 
+    cd .. && rm -rf "$1.dir"
+    if [ -f ./"$SZFILE" ]; then
+      rm "$1"
+    fi
+  else
+    _skip "$1 is not a zip file"
   fi
 }
 
@@ -45,18 +53,15 @@ _getbin
 while [ $# -ne 0 ]; do
   if [ -d "$1" ]; then
     cd "$1"
-      for ZIPFILE in *.zip; do
+      for ZIPFILE in ./*; do
         _zipify "$ZIPFILE"
       done
     cd "$PWD"
   elif [ -f "$1" ]; then
-    if $(file -i "$1" | grep -q 'application/zip'); then
-      cd $(dirname "$1")
-      _zipify $(basename "$1")
-      cd "$PWD"
-    else
-      _die "$1 is not a zip file"
-    fi
+    cd $(dirname "$1")
+    ZIPFILE=$(basename "$1")
+    _zipify "$ZIPFILE"
+    cd "$PWD"
   else
     _die "$1 does not exist"
   fi
